Permissions management back & front

+++ app/Http/Controllers/CityController.php
@@ -8,20 +8,23 @@ use App\Models\Zone;
 use App\Models\Stop;
 use App\Models\Role;
 use App\User;
+use Illuminate\Support\Facades\Auth;
 
 class CityController extends Controller {
 
     public function getAll(Request $request){
-        $user = User::find(1);
         return response()->json($user->getCities(), 200);
     }
 
     public function getOne(Request $request, City $city){
-        $user = User::find(1);
         return response()->json($city, 200);
     }
 
     public function getZones(Request $request, City $city) {
+        $user = Auth::user();
+        if( !$user->can('poi_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         $zones = Zone::where('city_id', $city->id)
             ->orderBy('name', 'asc')
             ->get();
@@ -29,10 +32,18 @@ class CityController extends Controller {
     }
 
     public function getZone(Request $request, City $city, Zone $zone) {
+        $user = Auth::user();
+        if( !$user->can('zone_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         return response()->json($zone, 200);
     }
 
     public function createZone(Request $request, City $city) {
+        $user = Auth::user();
+        if( !$user->can('zone_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         $zone = Zone::create([
             'name' => $request->name,
             'city_id' => $city->id,
@@ -41,6 +52,10 @@ class CityController extends Controller {
     }
 
     public function saveZone(Request $request, City $city, Zone $zone) {
+        $user = Auth::user();
+        if( !$user->can('zone_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         $zone->update([
             'name' => $request->name,
         ]);
@@ -48,15 +63,27 @@ class CityController extends Controller {
     }
 
     public function deleteZone(Request $request, City $city, Zone $zone) {
+        $user = Auth::user();
+        if( !$user->can('zone_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         Zone::destroy($zone->id);
         return response()->json(null, 204);
     }
 
     public function getGym(Request $request, City $city, Stop $stop) {
+        $user = Auth::user();
+        if( !$user->can('poi_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         return response()->json($stop, 200);
     }
 
     public function createGym(Request $request, City $city) {
+        $user = Auth::user();
+        if( !$user->can('poi_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         $gym = Stop::create([
             'name' => $request->name,
             'niantic_name' => $request->niantic_name,
@@ -72,6 +99,10 @@ class CityController extends Controller {
     }
 
     public function saveGym(Request $request, City $city, Stop $stop) {
+        $user = Auth::user();
+        if( !$user->can('poi_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         $stop->update([
             'name' => $request->name,
             'niantic_name' => $request->niantic_name,
@@ -98,6 +129,10 @@ class CityController extends Controller {
     }
 
     public function deleteGym(Request $request, City $city, Stop $stop) {
+        $user = Auth::user();
+        if( !$user->can('poi_edit', ['city_id' => $city->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         Zone::destroy($stop->id);
         return response()->json(null, 204);
     }
